+++ CHANGELOG.rst
@@ -1,6 +1,9 @@
-0.4.1
+0.5.0
 -----
 
+* Add support for running queries directly on tables with :func:`.sql_query`.
+* When creating tables, ``NOT NULL`` constraints will be created by default.
+* SQL create statements can now be generated without being executed with :func:`to_sql_create_statement`
 
 0.4.0 - December 19, 2016
 -------------------------